requirements.txt
llama-index==0.9.13
langchain==0.0.345
sentence-transformers
weaviate-client
pypdf
python-box